+++ package.json
@@ -17,6 +17,7 @@
 		"all": "run-s lint clean build gen schema test",
 		"check:build": "find . -type f -name \"*.d.ts\" -print | xargs grep \"packages/@markuplint\"",
 		"cli": "./packages/markuplint/bin/markuplint.mjs",
+		"create-rule": "./packages/@markuplint/create-rule/bin/create-rule.mjs",
 		"vscode:dev": "cd vscode; yarn --ignore-engines; yarn --ignore-engines vscode:dev",
 		"vscode:lint": "eslint --fix './vscode/**/*.ts'; yarn spellcheck; prettier --write './vscode/**/*{ts,tsx,js,cjs,mjs,json,rc,md}';",
 		"vscode:login": "cd vscode; yarn --ignore-engines vscode:login",
