--- packages/@markuplint/create-rule-helper/README.md
@@ -1,18 +1,0 @@
-# @markuplint/create-rule-helper
-
-[![npm version](https://badge.fury.io/js/%40markuplint%2Fcreate-rule-helper.svg)](https://www.npmjs.com/package/@markuplint/create-rule-helper)
-
-## Install
-
-[`markuplint`](https://www.npmjs.com/package/markuplint) package includes this package.
-
-<details>
-<summary>If you are installing purposely, how below:</summary>
-
-```shell
-$ npm install @markuplint/create-rule-helper
-
-$ yarn add @markuplint/create-rule-helper
-```
-
-</details>
