--- packages/@markuplint/create-rule-helper/src/index.ts
@@ -1,3 +1,0 @@
-export * from './create-rule-helper.js';
-export * from './types.js';
-export { isMarkuplintRepo } from './is-markuplint-repo.js';
