+++ packages/@markuplint/create-rule/README.md
@@ -1,0 +1,9 @@
+# @markuplint/create-rule
+
+[![npm version](https://badge.fury.io/js/%40markuplint%2Fcreate-rule.svg)](https://www.npmjs.com/package/@markuplint/create-rule)
+
+## Usage
+
+```shell
+$ npx @markuplint/create-rule <rule-name>
+```
