+++ packages/@markuplint/create-rule/bin/create-rule.mjs
@@ -1,0 +1,9 @@
+#!/usr/bin/env node
+
+import { createRule } from '../lib/cli.js';
+
+await createRule().catch(error => {
+	process.stderr.write(error + '\n');
+	process.exit(1);
+});
+process.exit(0);
