+++ packages/@markuplint/create-rule/scaffold/.eslintrc
