+++ packages/@markuplint/create-rule/scaffold/core/README.ja.md
