+++ packages/@markuplint/create-rule/scaffold/core/README.md
