+++ packages/@markuplint/create-rule/scaffold/core/index.spec.ts
