+++ packages/@markuplint/create-rule/scaffold/core/schema.json
