+++ packages/@markuplint/create-rule/scaffold/package/src/index.ts
