+++ packages/@markuplint/create-rule/scaffold/package/src/rules/__ruleName__.ts
