+++ packages/@markuplint/create-rule/scaffold/package/tsconfig.json
