+++ packages/@markuplint/create-rule/scaffold/project/rules/__ruleName__.spec.ts
