+++ packages/@markuplint/create-rule/src/install-scaffold.ts
