+++ packages/@markuplint/create-rule/src/read-package-json.ts
