+++ packages/@markuplint/create-rule/src/search-core-repository.ts
