+++ packages/markuplint/package.json
@@ -25,7 +25,6 @@
 		"clean": "tsc --build --clean"
 	},
 	"dependencies": {
-		"@markuplint/create-rule-helper": "4.0.0-alpha.6",
 		"@markuplint/cli-utils": "4.0.0-alpha.6",
 		"@markuplint/file-resolver": "4.0.0-alpha.6",
 		"@markuplint/html-parser": "4.0.0-alpha.6",
