+++ test/isolated-env/esm/package.json
@@ -12,7 +12,6 @@
 	"dependencies": {
 		"@markuplint/astro-parser": "file:../../../../markuplint/packages/@markuplint/astro-parser",
 		"@markuplint/config-presets": "file:../../../../markuplint/packages/@markuplint/config-presets",
-		"@markuplint/create-rule-helper": "file:../../../../markuplint/packages/@markuplint/create-rule-helper",
 		"@markuplint/cli-utils": "file:../../../../markuplint/packages/@markuplint/cli-utils",
 		"@markuplint/ejs-parser": "file:../../../../markuplint/packages/@markuplint/ejs-parser",
 		"@markuplint/erb-parser": "file:../../../../markuplint/packages/@markuplint/erb-parser",
