+++ test/isolated-env/ts-config-ts/package.json
@@ -13,7 +13,6 @@
 		"@types/node": "latest",
 		"@markuplint/astro-parser": "file:../../../../markuplint/packages/@markuplint/astro-parser",
 		"@markuplint/config-presets": "file:../../../../markuplint/packages/@markuplint/config-presets",
-		"@markuplint/create-rule-helper": "file:../../../../markuplint/packages/@markuplint/create-rule-helper",
 		"@markuplint/cli-utils": "file:../../../../markuplint/packages/@markuplint/cli-utils",
 		"@markuplint/ejs-parser": "file:../../../../markuplint/packages/@markuplint/ejs-parser",
 		"@markuplint/erb-parser": "file:../../../../markuplint/packages/@markuplint/erb-parser",
